Data_ingestion/requirements.txt
fastapi
uvicorn
openai
pinecone
beautifulsoup4
requests
python-dotenv
langchain-text-splitters
python-dateutil
numpy